+++ dice_backend/src/runtime/consts.rs
@@ -1,7 +1,7 @@
 /// Dice3 is a constant collection of the values a d3 can roll
 #[derive(Clone)]
 pub struct Dice3 {
-    arg: [i32; 3],
+    arg: [i8; 3],
 }
 impl Dice3 {
     /// build a new Dice3
@@ -10,7 +10,7 @@ impl Dice3 {
     }
 }
 impl IntoIterator for Dice3 {
-    type Item = i32;
+    type Item = i8;
     type IntoIter = Dice3Iter;
     fn into_iter(self) -> Dice3Iter {
         Dice3Iter {
@@ -24,15 +24,15 @@ impl IntoIterator for Dice3 {
 #[derive(Clone)]
 pub struct Dice3Iter {
     pos: usize,
-    arg: [i32; 3],
+    arg: [i8; 3],
 }
 impl Iterator for Dice3Iter {
-    type Item = i32;
-    fn next(&mut self) -> Option<i32> {
+    type Item = i8;
+    fn next(&mut self) -> Option<i8> {
         if self.pos >= 3 {
             None
         } else {
-            let x: i32 = self.arg[self.pos].clone();
+            let x: i8 = self.arg[self.pos].clone();
             self.pos += 1;
             Some(x)
         }
@@ -56,7 +56,7 @@ impl ::std::iter::ExactSizeIterator for Dice3Iter {
 /// Dice6 is a constant collection of what dice a d6 can roll
 #[derive(Clone)]
 pub struct Dice6 {
-    arg: [i32; 6],
+    arg: [i8; 6],
 }
 impl Dice6 {
     pub const fn new() -> Dice6 {
@@ -66,7 +66,7 @@ impl Dice6 {
     }
 }
 impl IntoIterator for Dice6 {
-    type Item = i32;
+    type Item = i8;
     type IntoIter = Dice6Iter;
     fn into_iter(self) -> Dice6Iter {
         Dice6Iter {
@@ -80,15 +80,15 @@ impl IntoIterator for Dice6 {
 #[derive(Clone)]
 pub struct Dice6Iter {
     pos: usize,
-    arg: [i32; 6],
+    arg: [i8; 6],
 }
 impl Iterator for Dice6Iter {
-    type Item = i32;
-    fn next(&mut self) -> Option<i32> {
+    type Item = i8;
+    fn next(&mut self) -> Option<i8> {
         if self.pos >= 6 {
             None
         } else {
-            let x: i32 = self.arg[self.pos].clone();
+            let x: i8 = self.arg[self.pos].clone();
             self.pos += 1;
             Some(x)
         }
